+++ packages/provider-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ai-sdk/provider-utils
 
+## 3.0.0-beta.7
+
+### Patch Changes
+
+- e7fcc86: feat (ai): introduce dynamic tools
+
 ## 3.0.0-beta.6
 
 ### Patch Changes